change_column

+++ db/migrate/20200812150409_add_column_to_users.rb
@@ -4,8 +4,8 @@ class AddColumnToUsers < ActiveRecord::Migration[5.2]
     add_column :users, :name, :string
     add_column :users, :gender, :string
     add_column :users, :user_image, :string
-    add_column :users, :height, :integer
-    add_column :users, :weight, :integer
-    add_column :users, :basal_metabolism, :integer
+    add_column :users, :height, :float
+    add_column :users, :weight, :float
+    add_column :users, :basal_metabolism, :float
   end
 end
